chore: небольшие правки в структуре файлов после ревью 11 спринта
