+++ ImageFeed/Auth/AuthViewController.swift
@@ -1,6 +1,5 @@
 
 import UIKit
-import ProgressHUD
 
 protocol AuthViewControllerDelegate: AnyObject {
     func didAuthenticate(_ vc: AuthViewController, code: String)
