+++ ImageFeed/Profile/ProfileResult.swift
@@ -1,0 +1,16 @@
+
+import UIKit
+
+struct ProfileResult: Codable {
+    let username: String?
+    let firstName: String?
+    let lastName: String?
+    let bio: String?
+    
+    enum CodingKeys: String, CodingKey {
+        case username = "username"
+        case firstName = "first_name"
+        case lastName = "last_name"
+        case bio = "bio"
+    }
+}
